rename adapt module and improve docs
