+++ src/adapt/adapt.ts
