+++ src/adapt/gro_adapter_generic_build.ts
@@ -1,6 +1,6 @@
 import {strip_trailing_slash} from '@feltcoop/felt/util/path.js';
 
-import type {Adapter} from 'src/adapt/adapter.js';
+import type {Adapter} from 'src/adapt/adapt.js';
 import type {Host_Target} from 'src/adapt/utils.js';
 import {copy_dist, ensure_nojekyll} from './utils.js';
 import {DIST_DIRNAME} from '../paths.js';
