+++ src/docs/plugin.md
@@ -8,7 +8,7 @@ Gro's builtin plugins:
 - [`gro_plugin_api_server`](../plugin/gro_plugin_api_server.ts)
 - [`gro_plugin_sveltekit_frontend `](../plugin/gro_plugin_sveltekit_frontend.ts)
 
-See [`config.plugin` in the config docs](config.md#plugin)
+Also see [`config.plugin` in the config docs](config.md#plugin)
 and usage in [the default config](../config/gro.config.default.ts).
 
 The implementation is at [`src/plugin/plugin.ts`](../plugin/plugin.ts) with more details.
